Removed commented code and migrated the index file to main plugin file
